+++ lib/src/entity/path.dart
@@ -10,4 +10,15 @@ class Path {
       path.add(PathPoint(element));
     }
   }
+
+  Map toJson() {
+    Map<String, dynamic> finalMap = {};
+    for (final coordinates in path) {
+      finalMap.putIfAbsent(path.indexOf(coordinates).toString(), () => {
+        coordinates.latitude,
+        coordinates.longitude
+      });
+    }
+    return finalMap;
+  }
 }